Build design stage classes without knowing the type
